Fix range fetch for Android Youtube app
Sometimes, Youtube app will try to do range fetch for a video. This action relys
on the "Content-Range" and "Content-Length" in the http header to identify
the next part of video to fetch (actually this logic exists in StageFright
media player library).

However, GoAgent choose to not send back "Content-Range" and "Content-Length" for
range fetch requests. This causes Youtube fetching wrong part of video again and
again. At last, the playback becomes jitter and stutter.

This patch aims to fix the problem, but without enough testing, I cannot make
sure that there exists no side effects for other platform or software.